% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainhibtaoga.R
\name{AlooParatha}
\alias{AlooParatha}
\title{Create Histogram,boxplot and numeric summary'}
\usage{
AlooParatha(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Create Histogram,boxplot and numeric summary'
}
